only account for header height if Window has a title
This logic is in GroupLayout, but absent from some other layouts. Without this check, there can be padding if, for instance, one uses a GridLayout with a Popup.

Note: I have not tested all layouts in all configurations. For instance, I'm not sure if the behaviour makes sense for Orientation::Vertical.