Fix tab controls (in VS 2015 release mode)
The eigen based computations in the TabHeader use auto to store Eigen's
ArrayWrapper objects which, for some reason, isn't kept alive (is it an rvalue?)
In Microsoft Visual Studio 2015 this leads to the arrow buttons being in the
wrong position, but only when compiling in release mode.
The issue is visible even in Example1.cpp.